# docs/swagger.yaml
basePath: /api/v1
definitions:
  main.UserCommandDto:
    properties:
      age:
        type: integer
      password:
        type: string
      username:
        type: string
    required:
    - age
    - password
    - username
    type: object
  models.User:
    properties:
      created_at:
        type: string
      family:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  userdto.UserCommandDto:
    properties:
      family:
        type: string
      name:
        type: string
    required:
    - family
    - name
    type: object
host: localhost:8000
info:
  contact: {}
  description: Minimal Gin + Swagger example
  title: Example API
  version: "1.0"
paths:
  /user/add:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/userdto.UserCommandDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add new user
      tags:
      - User
  /user/delete/{id}:
    delete:
      description: Delete a user by ID
      parameters:
      - description: ID of target user
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete user
      tags:
      - User
  /user/list:
    get:
      description: Get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/models.User'
              type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user list
      tags:
      - User
  /user/userdetail:
    post:
      consumes:
      - application/json
      description: Echo user
      parameters:
      - description: User information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.UserCommandDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.UserCommandDto'
      summary: Test handler
      tags:
      - User
schemes:
- http
swagger: "2.0"